Remove compiler warnings.
git-svn-id: https://randomcoder.org/repos/public/randomcoder-proxy-client/trunk@988 5bee6cb3-3d18-0410-8c93-a642edd49b48

+++ src/main/java/com/randomcoder/proxy/client/ProxyInputStream.java
@@ -137,7 +137,7 @@ public class ProxyInputStream extends InputStream
 		int result = inputStream.read(b, off, len);
 		
 		if (result > 0 && listener != null)
-			listener.dataReceived(null, (long) result);
+			listener.dataReceived(null, result);
 		
 		return result;
 	}
@@ -148,7 +148,7 @@ public class ProxyInputStream extends InputStream
 		int result = inputStream.read(b);
 		
 		if (result > 0 && listener != null)
-			listener.dataReceived(null, (long) result);
+			listener.dataReceived(null, result);
 		
 		return result;
 	}
